Cleaned up verilog to gds file by moving additional functions to utilities file - still no failed routes
